Debugger: map signal numbers
Remote debugging: The signal numbers on the OS running Debugger and the
OS running the debugged team might not be identical. So we need a
mapping to stable numbers internal to Debugger.

The same holds for pretty much all enums and structs. A lot of them
already have the required OS independent representation, but some don't.