Terminal: new feature: custom hyperlink commands
One can define additional commands that appear in the hyper link menu
(Command + right click on file name). There's currently no GUI to
define them, just the driver settings file
~/config/settings/Terminal/HyperLinkCommands. It may contain any number
of entries like this:

CustomCommand {
	label	"Git Diff"
	command	"git diff %f | lpe --type diff"
}

"label" specifies the label of the menu item, "command" the command
line to be executed when the menu item is selected. Supported
placeholders are '%f', the file path for which the menu was invoked
and '%d' the current foreground process' working directory, both
shell-escaped.

There's a lot of room for improvements:
* support for URLs (currently only file links),
* better support for file links with line/column number (separate
  placeholders for file path and line/column),
* add menu items depending on the context in the Terminal (e.g.
  have a "Git Diff" item only when a "git status" precedes the
  occurrence of the file link).