AutoDeleter: c'tor/assignment operator safety
* Make all single parameter c'tors explicit.
* Hide copy constructor and assignment operator.

This prevents unintended behavior. Assigning a deletable to an
AutoDeleter would compile fine, but would not behave as expected: A
temporary AutoDeleter would be created, the default assignment operator
would be used to assign to the existing AutoDeleter, and the temporary
one would be destroyed, deleting the deletable while the existing
AutoDeleter would still refer to it.