Debugger: StreamMessenger: track in which queue a message is
Since we use DoublyLinkedLists, trying to remove a message that isn't
in the list isn't safe. Since we don't add a message to a queue when
there's already a listener waiting, fReceivedMessages.RemoveMessage()
may be invoked for a message that hasn't been added to
fReceivedMessages. Message now has bool members for both channel and
global queue to indicate whether it is in the respective queue.